src/app/core/components/user-panel/user-panel.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { UserPanelComponent } from './user-panel.component';

describe('UserPanelComponent', () => {
  let component: UserPanelComponent;
  let fixture: ComponentFixture<UserPanelComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ UserPanelComponent ],
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(UserPanelComponent);
    component = fixture.componentInstance;

    component.user = {
      id: '1',
      name: {
        first: 'User',
        last: 'User Lastname',
      },
    };
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should get user', () => {
    expect(component.id).toBeTruthy();
    expect(component.name.first).toBeTruthy();
    expect(component.name.last).toBeTruthy();
  });
});
